Changes AMAxisInfo and AMMeasurementInfo to use long instead of int for axis dimensions. (Consistent with increased range of AMnDIndex).

+++ source/dataman/AMAxisInfo.cpp
@@ -22,7 +22,7 @@ along with Acquaman.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "dataman/AMNumber.h"
 
-AMAxisInfo::AMAxisInfo(const QString& mName, int mSize, const QString& mDescription, const QString& mUnits)
+AMAxisInfo::AMAxisInfo(const QString& mName, long mSize, const QString& mDescription, const QString& mUnits)
 {
 	name = mName;
 	size = mSize;
